wdk2023: store adsp firmware in disabled form on the rootfs
to make it possible to boot from USB Type-C the firmware needs to be in the initramfs already. However, if adsp is loaded there will
be a short interruption on VBUS. Ref:
https://gitlab.com/postmarketOS/pmaports/-/blob/master/device/testing/firmware-lenovo-yoga-5g/APKBUILD?ref_type=heads#L44-46
For some odd reasons not compleately understood, adsp will tried to be loaded from rootfs even if it is already loaded from
initramfs. If that happens and you're booted from an USB Type-C, it will render the rootfs unreadable, essentially preventing
to bootup the system. It can be avoided if the adsp firmware is available from initramfs only. Therefore store the adsp file
only with .disabled suffix.

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>